refactor(fetch): client side fetching for films, search, actors, user page
