+++ src/app/(search)/tv/search/page.js
@@ -1,9 +1,5 @@
-import Search from "@/components/Search/";
-import Filters from "@/components/Search/Filter";
-import { axios } from "@/lib/axios";
-import { tvGenres } from "@/data/tv-genres";
-import { languages } from "@/data/languages";
 import { siteConfig } from "@/config/site";
+import ClientTvSearch from "@/components/Search/ClientTvSearch";
 
 export async function generateMetadata() {
   return {
@@ -15,49 +11,7 @@ export async function generateMetadata() {
   };
 }
 
-export default async function page() {
-  const [{ results: fetchMinYear }, { results: fetchMaxYear }] =
-    await Promise.all([
-      axios
-        .get(`/discover/tv`, {
-          params: {
-            sort_by: "first_air_date.asc",
-          },
-        })
-        .then(({ data }) => data),
-
-      axios
-        .get(`/discover/tv`, {
-          params: {
-            sort_by: "first_air_date.desc",
-          },
-        })
-        .then(({ data }) => data),
-    ]);
-
-  const defaultMaxYear = new Date().getFullYear() + 1;
-  const minYear = new Date(fetchMinYear[0].first_air_date).getFullYear();
-  const maxYear = new Date(fetchMaxYear[0].first_air_date).getFullYear();
-
-  return (
-    <div className={`flex gap-4 lg:px-4`}>
-      <h1 className="sr-only">Search TV Shows</h1>
-
-      <Filters
-        type={"tv"}
-        genresData={tvGenres}
-        minYear={minYear}
-        maxYear={maxYear}
-        languagesData={languages}
-      />
-
-      <Search
-        type={`tv`}
-        genresData={tvGenres}
-        languagesData={languages}
-        minYear={minYear}
-        maxYear={maxYear > defaultMaxYear ? defaultMaxYear : maxYear}
-      />
-    </div>
-  );
+export default function page() {
+  return <ClientTvSearch />;
 }
+
